Update framework to handle Overrides on PS 1.4 + Make module compliant with PS 1.4

+++ froggy/FroggyModule.php
@@ -57,7 +57,10 @@ class FroggyModule extends Module
 
 		// If PS version is lower than 1.5, call backward script
 		if (version_compare(_PS_VERSION_, '1.5') < 0)
+		{
 			require(dirname(__FILE__).'/FroggyBackward.php');
+			require_once(dirname(__FILE__).'/FroggyOverride.php');
+		}
 
 		// Define local path if not exists (1.4 compatibility)
 		if (!isset($this->local_path))
@@ -69,6 +72,10 @@ class FroggyModule extends Module
 			smartyRegisterFunction($this->context->smarty, 'function', 'FroggyGetAdminLink', 'FroggyGetAdminLink');
 			$this->context->smarty_methods['FroggyGetAdminLink'] = true;
 		}
+
+		// Define module configuration url
+		if (isset($this->context->employee->id))
+			$this->configuration_url = 'index.php?controller=AdminModules&token='.Tools::getAdminTokenLite('AdminModules').'&configure='.$this->name.'&module_name='.$this->name;
 	}
 
 	/**
@@ -138,6 +145,9 @@ class FroggyModule extends Module
 	{
 		if (parent::install())
 		{
+			if (class_exists('FroggyOverride') && !$this->installFroggyOverrides())
+				return false;
+
 			if (!$this->registerDefinitionsHooks())
 				return false;
 
@@ -149,6 +159,7 @@ class FroggyModule extends Module
 
 			if (!$this->runDefinitionsSql())
 				return false;
+
 			return true;
 		}
 
@@ -164,15 +175,50 @@ class FroggyModule extends Module
 	{
 		if (parent::uninstall())
 		{
+			if (class_exists('FroggyOverride') && !$this->uninstallFroggyOverrides())
+				return false;
+
 			if (!$this->deleteConfigurations())
 				return false;
+
 			if (!$this->deleteModuleControllers())
 				return false;
+
 			if (!$this->runDefinitionsSql('uninstall'))
 				return false;
+
 			return true;
 		}
 		return false;
+	}
+
+	/**
+	 * Install overrides files for the module
+	 *
+	 * @return bool
+	 */
+	public function installFroggyOverrides()
+	{
+		$fo = new FroggyOverride($this->name);
+		try {
+			$fo->installOverrides();
+		} catch (Exception $e) {
+			$this->_errors[] = sprintf(Tools::displayError('Unable to install override: %s'), $e->getMessage());
+			$fo->uninstallOverrides();
+			return false;
+		}
+		return true;
+	}
+
+	/**
+	 * Uninstall overrides files for the module
+	 *
+	 * @return bool
+	 */
+	public function uninstallFroggyOverrides()
+	{
+		$fo = new FroggyOverride($this->name);
+		return $fo->uninstallOverrides();
 	}
 
 	/**
